refactor: move setupGlobalPreloadHandler to plugin
