+++ packages/demo/src/routes/layout.tsx
@@ -1,6 +1,5 @@
 import BarOfProgress from "@badrap/bar-of-progress";
 import { getTheme, setTheme } from "@hiogawa/theme-script";
-import { getRouteDependencies } from "@hiogawa/vite-glob-routes/dist/react-router/client";
 import React from "react";
 import { Toaster } from "react-hot-toast";
 import {
@@ -72,8 +71,6 @@ function PageInner() {
 }
 
 function Header() {
-  useGlobalPagePrefetchHandler();
-
   return (
     <header className="top-0 sticky antd-body flex items-center p-2 px-4 gap-3 shadow-md shadow-black/[0.05] dark:shadow-black/[0.7] z-1">
       <div>Example</div>
@@ -116,56 +113,6 @@ const ROUTES = [
   "/error",
   "/session",
 ];
-
-//
-// link prefetching
-//
-
-// cf. https://github.com/remix-run/remix/blob/9ae3cee0e81ccb7259d6103df490b019e8c2fd94/packages/remix-react/components.tsx#L479
-function useGlobalPagePrefetchHandler() {
-  React.useEffect(() => {
-    // TODO: not sure perf impact of global "mouseover"
-    function handler(e: MouseEvent | TouchEvent) {
-      if (e.target instanceof HTMLAnchorElement) {
-        const dataPrefetch = e.target.getAttribute("data-prefetch");
-        if (dataPrefetch) {
-          // e.target.href always full url?
-          injectPagePrefetchLinks(e.target.href);
-        }
-      }
-    }
-
-    document.addEventListener("mouseover", handler);
-    document.addEventListener("touchstart", handler);
-    return () => {
-      document.removeEventListener("mouseover", handler);
-      document.removeEventListener("touchstart", handler);
-    };
-  }, []);
-}
-
-// TODO: memoize?
-function injectPagePrefetchLinks(href: string) {
-  const url = new URL(href, window.location.href);
-
-  // TODO: prefetch external links?
-  if (url.host !== window.location.host) {
-    return;
-  }
-
-  // resolve page dependencies
-  const links = getRouteDependencies(url.pathname);
-  for (const href of links.js) {
-    // TODO: escapeHtml
-    const found = document.body.querySelector(`link[href="${href}"]`);
-    if (!found) {
-      const el = document.createElement("link");
-      el.setAttribute("rel", "modulepreload");
-      el.setAttribute("href", href);
-      document.body.appendChild(el);
-    }
-  }
-}
 
 //
 // ThemeSelect
